move data transform logic to route handler
